feat: create controller, repository and service

+++ src/main/java/com/billing/finalproject/entity/Client.java
@@ -11,7 +11,7 @@ import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 
 @Entity
-@Table (name = "clients")
+@Table(name = "clients")
 public class Client {
     public Client() {
         super();
@@ -19,45 +19,53 @@ public class Client {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column (name = "id")
-    private int id;
-    @Column (name = "name")
+    @Column(name = "id")
+    private Long id;
+    @Column(name = "name")
     private String name;
-    @Column (name = "lastname")
+    @Column(name = "lastname")
     private String lastname;
-    @Column (name = "docnumber")
+    @Column(name = "docnumber")
     private String docnumber;
 
     @OneToMany(mappedBy = "client")
     private List<Invoice> invoices;
 
-    public Client (int id, String name, String lastname, String docnumber) {
+    public Client(Long id, String name, String lastname, String docnumber) {
         this.id = id;
         this.name = name;
         this.lastname = lastname;
         this.docnumber = docnumber;
     }
-    public int getId() {
+
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+
+    public void setId(Long id) {
         this.id = id;
     }
+
     public String getName() {
         return name;
     }
+
     public void setName(String name) {
         this.name = name;
     }
+
     public String getLastname() {
         return lastname;
     }
+
     public void setLastname(String lastname) {
         this.lastname = lastname;
     }
+
     public String getDocnumber() {
         return docnumber;
     }
+
     public void setDocnumber(String docnumber) {
         this.docnumber = docnumber;
     }
