+++ src/main/java/com/billing/finalproject/entity/Product.java
@@ -10,67 +10,75 @@ import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.GenerationType;
 
-
 @Entity
-@Table (name = "products")
+@Table(name = "products")
 public class Product {
     public Product() {
         super();
     }
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column (name = "id")
-    private int id;
-    @Column (name = "description")
+    @Column(name = "id")
+    private Long id;
+    @Column(name = "description")
     private String description;
-    @Column (name = "code")
+    @Column(name = "code")
     private String code;
-    @Column (name = "stock")
+    @Column(name = "stock")
     private int stock;
-    @Column (name = "price")
+    @Column(name = "price")
     private double price;
 
     @OneToMany(mappedBy = "product")
     private List<InvoiceDetails> invoiceDetails;
 
-    public Product (int id, String description, String code, int stock, double price) {
+    public Product(Long id, String description, String code, int stock, double price) {
         this.id = id;
         this.description = description;
         this.code = code;
         this.stock = stock;
         this.price = price;
     }
-    
-    public int getId() {
+
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+
+    public void setId(Long id) {
         this.id = id;
     }
+
     public String getDescription() {
         return description;
     }
+
     public void setDescription(String description) {
         this.description = description;
     }
+
     public String getCode() {
         return code;
     }
+
     public void setCode(String code) {
         this.code = code;
     }
+
     public int getStock() {
         return stock;
     }
+
     public void setStock(int stock) {
         this.stock = stock;
     }
+
     public double getPrice() {
         return price;
     }
+
     public void setPrice(double price) {
         this.price = price;
     }
 
-    
 }
